Phase 2: move drawing.py/measurement.py into dense_evolution.utils/
First subpackage of the library-side refactor (tracking issue #76), chosen
as the least-dependent leaf module per the migration order in prog.txt.

Real content now lives at dense_evolution.utils.drawing/.measurement.
Backward-compatible shims left at the old top-level paths
(dense_evolution/drawing.py, dense_evolution/measurement.py) so external
consumers importing dense_evolution.drawing/.measurement directly (e.g.
Dense-Evolution-Discovery) keep working unchanged -- verified the shim and
the new canonical path resolve to the literal same function object.

pyproject.toml's packages list and the mkdocstrings directives in
docs/api/drawing.md and docs/api/measurement.md updated to the new
canonical location.

Verified: pip install -e . + import smoke test (new path, old shim path,
identity check) + isolated test_mcp_server.py/test_local_site_server.py
(89/89 passed) to rule out the 6 failures seen in a full-suite run as
resource contention (confirmed pre-existing/unrelated, not caused by this
change -- reproduces even without touching mcp_server/local_site code).
